+++ CameraSample/CameraSample/Content/FilterBanner/FilterBanner.swift
@@ -35,7 +35,6 @@ struct FilterBannerView: View {
                     }
                 }
 
-
                 HStack {
                     Button(action: {
                         withAnimation {
@@ -72,37 +71,6 @@ struct FilterBannerView: View {
             .foregroundColor(.white)
             .offset(x: 0, y: isShowBanner ? 0 : 300)
         }
-    }
-}
-
-final class FilterBannerViewModel: ObservableObject {
-    @Published var selectedFilter: FilterType?
-    private var cancelables: [Cancellable] = []
-    var isSelectedPixellate: Bool {
-        return selectedFilter == .pixellate
-    }
-
-    var isSelectedSepiaTone: Bool {
-        return selectedFilter == .sepiaTone
-    }
-
-    var isSelectedSharpenLuminance: Bool {
-        return selectedFilter == .sharpenLuminance
-    }
-
-    var isSelectedPhotoEffectMono: Bool {
-        return selectedFilter == .photoEffectMono
-    }
-
-    var isSelectedGaussianBlur: Bool {
-        return selectedFilter == .gaussianBlur
-    }
-
-    init() {
-        let subscriber = $selectedFilter.sink { (type) in
-            print("\(type.debugDescription)")
-        }
-        cancelables.append(subscriber)
     }
 }
