+++ CameraSample/CameraSample/Content/ImagePicker/ImagePicker.swift
