+++ CameraSample/CameraSample/Model/FilterType.swift
@@ -7,6 +7,8 @@
 //
 
 import UIKit
+import CoreImage
+import CoreImage.CIFilterBuiltins
 
 enum FilterType: String, CaseIterable {
     case pixellate = "モザイク"
@@ -15,7 +17,7 @@ enum FilterType: String, CaseIterable {
     case photoEffectMono = "モノクロ"
     case gaussianBlur = "ブラー"
 
-//https://www.hackingwithswift.com/books/ios-swiftui/integrating-core-image-with-swiftui
+    //https://www.hackingwithswift.com/books/ios-swiftui/integrating-core-image-with-swiftui
     private func makeFilter(inputImage: CIImage?) -> CIFilterProtocol {
 
         switch self {
